Remove Foundation application from translation documentation (#179)
Designation of this form in the translation documentation was an aberration to community convention. I hope to take this application process in a different direction, an evolution of its previous situation.

Um... I mean... This removes the UBports Foundation application from the page about translations. Hopefully this will reduce the number of people who accidentally stumble upon the form.