User manual: SD card section updatet pr. review
Removed section about automount being needed, it will issue a warning if the card is not mounted
Clarified how to use deep links
Expanded upon the "feature" of showexec and why it is a problem. I don't know if it's suddenly a bit long winded and explanatory for a user manual?