Change "Porting information" to "Halium porting"
To make way for new information about building ubp-5.1, the porting docs
needed to be reorganized a bit. I'm starting by changing "Porting
information" to "Halium porting" to make it a little easier to
understand what the page is for.

I've also added more information about what this page is or is not for.