Update bug reporting document
There was some confusion about the steps to get logs, and the steps
don't always work. Sometimes 'dmesg' requires 'sudo' and 'logcat' always
does, causing problems for 'adb shell'. Since this is the case,
I moved the log section to after posting the bug. The triager will know
the best logs to get and the best way to get them anyway.