Redirect old porting pages to new halium_7-1 pages
The merge of #342[1] brought some awesome new content, but it also
removed some pages. We have links to these pages elsewhere (other
documentation, our website, blogs) and it would be a shame to send those
to a 404.

To fix this, set up Rediraffe to check for files removed in the current
branch. This will warn us in CI output when a file is removed without a
redirect. By adding the 'rediraffecheckdiff' build at the end of
build.sh, people can easily check their own docs.

Rediraffe also allows *creating* redirects from the newly nonexistent
pages, so we'll start with the recently-removed porting pages in
redirects.txt.

[1] https://github.com/ubports/docs.ubports.com/pull/342